chore: refactor backend actions and changes
This aims at give better readability and maintainability to backend code.
