+++ backend/lib/edgehog/containers/container/deployment/changes/maybe_notify_upwards.ex
@@ -18,7 +18,7 @@
 # SPDX-License-Identifier: Apache-2.0
 #
 
-defmodule Edgehog.Containers.Container.Changes.MaybeNotifyUpwards do
+defmodule Edgehog.Containers.Container.Deployment.Changes.MaybeNotifyUpwards do
   @moduledoc """
   If the container deployment is ready, notify the associated deployments.
   """
