+++ backend/lib/edgehog/containers/image_credentials/image_credentials.ex
